# service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# CrossWave Xmatch API\n\nAPI for astronomical cross-matching and catalog queries developed by [ALeRCE](https://alerce.science).\n\n## Overview\n\nCrossWave provides fast cone search and metadata retrieval across multiple astronomical catalogs. The service is optimized for high-throughput queries using HEALPix spatial indexing.\n\n### Key Features\n\n- **Cone Search**: Find objects within a radius of given celestial coordinates\n- **Bulk Operations**: Process multiple queries in a single request\n- **Metadata Retrieval**: Get detailed catalog information for specific objects\n- **Lightcurves**: Retrieve time-series photometry data\n\n## Coordinates\n\nAll coordinates use the **J2000 equatorial coordinate system**:\n\n| Parameter | Range | Unit | Description |\n|-----------|-------|------|-------------|\n| `ra` | 0 to 360 | degrees | Right Ascension |\n| `dec` | -90 to +90 | degrees | Declination |\n| `radius` | \u003e 0 | degrees | Search radius |\n\n## Available Catalogs\n\n| Catalog | Description |\n|---------|-------------|\n| `all` | Search across all available catalogs |\n| `allwise` | WISE All-Sky Data Release |\n\n## Response Codes\n\n| Code | Description |\n|------|-------------|\n| `200` | Success - results found |\n| `204` | Success - no results found |\n| `400` | Bad Request - invalid parameters |\n| `500` | Internal Server Error |\n\n## Example Usage\n\n### Single Cone Search\n\nSearch for objects within 0.01 degrees of RA=180.5, Dec=-45.0:\n\n```\nGET /v1/conesearch?ra=180.5\u0026dec=-45.0\u0026radius=0.01\u0026catalog=allwise\n```\n\n### Bulk Cone Search\n\nSearch multiple positions in a single request:\n\n```json\nPOST /v1/bulk-conesearch\nContent-Type: application/json\n\n{\n  \"ra\": [180.5, 181.2, 182.0],\n  \"dec\": [-45.0, -45.5, -46.0],\n  \"radius\": 0.01,\n  \"catalog\": \"allwise\",\n  \"nneighbor\": 1\n}\n```\n\n### Get Metadata by ID\n\nRetrieve detailed catalog data for a specific object:\n\n```\nGET /v1/metadata?id=J120000.00-450000.0\u0026catalog=allwise\n```\n\n### Get Lightcurve\n\nRetrieve time-series photometry for coordinates:\n\n```\nGET /v1/lightcurve?ra=180.5\u0026dec=-45.0\u0026radius=0.01\n```\n\n## Error Handling\n\nAll validation errors return a JSON object with the following structure:\n\n```json\n{\n  \"field\": \"ra\",\n  \"reason\": \"value out of range\",\n  \"value\": \"400.0\"\n}\n```\n\n## Rate Limits\n\n- Bulk endpoints accept up to **1000 coordinates** per request\n- Results are processed in parallel for optimal performance\n",
        "title": "CrossWave HTTP API",
        "contact": {
            "name": "Diego Rodriguez Mancini",
            "email": "diegorodriguezmancini@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/bulk-conesearch": {
            "post": {
                "description": "Search for objects in a given region using multiple RA/Dec coordinates and a single radius.\nAll coordinate pairs are searched in parallel for optimal performance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conesearch"
                ],
                "summary": "Bulk cone search for objects",
                "operationId": "bulk-conesearch",
                "parameters": [
                    {
                        "description": "Bulk conesearch request with arrays of coordinates",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BulkConesearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found objects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.Mastercat"
                            }
                        }
                    },
                    "204": {
                        "description": "No objects found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/conesearch.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bulk-metadata": {
            "post": {
                "description": "Retrieve detailed catalog metadata for multiple astronomical objects by their identifiers.\nAll IDs are queried in parallel for optimal performance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Bulk get metadata by object IDs",
                "operationId": "bulk-metadata",
                "parameters": [
                    {
                        "description": "Bulk metadata request with list of object IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BulkMetadataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Objects metadata",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.Allwise"
                            }
                        }
                    },
                    "204": {
                        "description": "No objects found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/metadata.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/conesearch": {
            "get": {
                "description": "Search for astronomical objects within a specified radius of given celestial coordinates (RA/Dec).\nReturns matching objects from the specified catalog.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "conesearch"
                ],
                "summary": "Cone search for objects",
                "operationId": "conesearch",
                "parameters": [
                    {
                        "maximum": 360,
                        "minimum": 0,
                        "type": "number",
                        "example": 180.5,
                        "description": "Right Ascension (J2000) in degrees",
                        "name": "ra",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 90,
                        "minimum": -90,
                        "type": "number",
                        "example": -45,
                        "description": "Declination (J2000) in degrees",
                        "name": "dec",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "number",
                        "example": 0.01,
                        "description": "Search radius in degrees",
                        "name": "radius",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Catalog to search in",
                        "name": "catalog",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Maximum number of neighbors to return",
                        "name": "nneighbor",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include full metadata in response",
                        "name": "getMetadata",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found objects",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.Mastercat"
                            }
                        }
                    },
                    "204": {
                        "description": "No objects found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/conesearch.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lightcurve": {
            "get": {
                "description": "Retrieve time-series photometry data (lightcurve) for astronomical objects at the specified coordinates.\nReturns detections, non-detections, and forced photometry measurements.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lightcurve"
                ],
                "summary": "Get lightcurve data",
                "operationId": "lightcurve",
                "parameters": [
                    {
                        "maximum": 360,
                        "minimum": 0,
                        "type": "number",
                        "example": 180.5,
                        "description": "Right Ascension (J2000) in degrees",
                        "name": "ra",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 90,
                        "minimum": -90,
                        "type": "number",
                        "example": -45,
                        "description": "Declination (J2000) in degrees",
                        "name": "dec",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "number",
                        "example": 5,
                        "description": "Search radius in arcseconds",
                        "name": "radius",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Number of neighbors to return",
                        "name": "nneighbor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lightcurve data",
                        "schema": {
                            "$ref": "#/definitions/lightcurve.Lightcurve"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/metadata": {
            "get": {
                "description": "Retrieve detailed catalog metadata for a specific astronomical object by its identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Get metadata by object ID",
                "operationId": "metadata",
                "parameters": [
                    {
                        "type": "string",
                        "example": "J120000.00-450000.0",
                        "description": "Object identifier",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "allwise",
                        "description": "Catalog to search in",
                        "name": "catalog",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object metadata",
                        "schema": {
                            "$ref": "#/definitions/repository.Allwise"
                        }
                    },
                    "204": {
                        "description": "Object not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/metadata.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BulkConesearchRequest": {
            "type": "object",
            "properties": {
                "catalog": {
                    "description": "Catalog name to search in (default: all)",
                    "type": "string",
                    "example": "allwise"
                },
                "dec": {
                    "description": "Declination values in degrees (-90 to 90)",
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        -45,
                        -45.5
                    ]
                },
                "nneighbor": {
                    "description": "Number of neighbors to return per coordinate (default: 1)",
                    "type": "integer",
                    "example": 1
                },
                "ra": {
                    "description": "Right ascension values in degrees (0-360)",
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        180.5,
                        181.2
                    ]
                },
                "radius": {
                    "description": "Search radius in degrees",
                    "type": "number",
                    "example": 0.01
                }
            }
        },
        "api.BulkMetadataRequest": {
            "type": "object",
            "properties": {
                "catalog": {
                    "description": "Catalog to search in",
                    "type": "string",
                    "example": "allwise"
                },
                "ids": {
                    "description": "List of object identifiers to search for",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "id1",
                        "id2"
                    ]
                }
            }
        },
        "conesearch.ValidationError": {
            "type": "object",
            "properties": {
                "errValue": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "lightcurve.Lightcurve": {
            "type": "object",
            "properties": {
                "detections": {
                    "type": "array",
                    "items": {}
                },
                "forced_photometry": {
                    "type": "array",
                    "items": {}
                },
                "non_detections": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "metadata.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "repository.Allwise": {
            "type": "object",
            "properties": {
                "cntr": {
                    "type": "integer"
                },
                "h_m_2mass": {
                    "type": "number"
                },
                "h_msig_2mass": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "j_m_2mass": {
                    "type": "number"
                },
                "j_msig_2mass": {
                    "type": "number"
                },
                "k_m_2mass": {
                    "type": "number"
                },
                "k_msig_2mass": {
                    "type": "number"
                },
                "w1mpro": {
                    "type": "number"
                },
                "w1sigmpro": {
                    "type": "number"
                },
                "w2mpro": {
                    "type": "number"
                },
                "w2sigmpro": {
                    "type": "number"
                },
                "w3mpro": {
                    "type": "number"
                },
                "w3sigmpro": {
                    "type": "number"
                },
                "w4mpro": {
                    "type": "number"
                },
                "w4sigmpro": {
                    "type": "number"
                }
            }
        },
        "repository.Mastercat": {
            "type": "object",
            "properties": {
                "cat": {
                    "type": "string"
                },
                "dec": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "ipix": {
                    "type": "integer"
                },
                "ra": {
                    "type": "number"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Search for astronomical objects by celestial coordinates (RA/Dec)",
            "name": "conesearch"
        },
        {
            "description": "Retrieve detailed catalog information for specific objects",
            "name": "metadata"
        },
        {
            "description": "Get time-series photometry data for objects",
            "name": "lightcurve"
        }
    ],
    "externalDocs": {
        "description": "ALeRCE Documentation",
        "url": "https://alerce.science"
    }
}